full executable py file
the code for parts have been collected into this py file, simply download this along with the data.csv file into a workspace and they would execute properly. if anything, do be sure to comment out the other code for the other analysis questions, e.g. if looking to run the code for problem 4, comment out everything below the dashed line in order to make sure the code for #4 is the only thing running.